requirements.txt
playwright>=1.40.0
pyyaml>=6.0
pydantic>=2.0
loguru>=0.7.0
aiohttp>=3.9.0
camoufox>=0.4.11
